Move all XML types out of xquery and into xml package.
